main.c/*.py: added a help thing to main and *.py when given no arguments main.c: added resizable window

+++ sdl_radix_compare.c
@@ -1,7 +1,8 @@
+#include <stdint.h>
 #include <SDL2/SDL_image.h>
 
 #include "sdl_radix_compare.h"
 
 int sdl_radix_compare(void *a, void *b) {
-	return *(unsigned int *)a & ((SDL_Rect *)b)->h;
+	return *(uint64_t *)a & (uint64_t)((SDL_Rect *)b)->h;
 }
